[rpc] catch exceptions in ProcessGroupAgent::enqueueSend and report them.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31023

Adds support to catch exceptions in ProcessGroupAgent::enqueueSend and
report them in the future by marking the future as completed with an exception
indicating the error. An example of when this could happen is if the receiving
side aborts when the sender is sending the message, previously, we would hang
until the timeout is hit, and the original exception would be lost.
ghstack-source-id: 96373664

Differential Revision: [D18901981](https://our.internmc.facebook.com/intern/diff/D18901981/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D18901981/)!