Add type annotation for bias in _ConvNd
Summary:
Currently Tensor bias is registered as parameter and None bias is registered as attribute.
We need the type annotation because when we try to fold ConvBn in graph mode quantization we'll
remove the None bias attribute and add a Tensor bias attribute, without type annotation the
bias Value in the graph  will be marked with different type in these two cases, so we have rewrite the
graph to change the type as well in that case. But with type annotation we don't need to modify the graph
since both cases the bias value will have type `Tensor?`

Test Plan:
.

Reviewers:
suo

Subscribers:

Tasks:

Tags:

ghstack-source-id: bf2a355ef17d87d55215bbe0043b0bc048003f72
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32885