[Lite Trainer] Add necessary registrations for MNIST model
Because of the special treatment of operator names for lite interpreter, all the operators used in lite interpreter are still prepended by "_". Add the necessary registrations for MNIST model. All the ops with autograd capability are included in torch_mobile_train. After rebase the selective build from D19649074 can be utilized to strip the unused ops.

Note that this diff is for feasibility test. The training accuracy are not covered in the test.

Differential Revision: [D19743201](https://our.internmc.facebook.com/intern/diff/D19743201/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D19743201/)!

ghstack-source-id: 97780066
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33717