Use clog from cpuinfo/deps instead of downloading
Summary:
XNNPACK was downloading clog because we weren't setting CLOG_SOURCE_DIR.
Actually, it was downloading cpuinfo and pointing to the copy of clog
within that.  So let's just point to the copy of clog within the cpuinfo
submodule we already have.

Test Plan:
Ran cmake and didn't see any downloading.
Verified that our clog is the same as the one that was being downloaded
with `diff -Naur`.

ghstack-source-id: c5025e0a4edc49dd92592e20052b8aedd8a5625a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33947